fix: extract pag-table to shared folder
