+++ apps/web/src/features/spaces/components/SpaceAddressBook/PendingRequestsTable.tsx
@@ -20,7 +20,7 @@ import { showNotification } from '@/store/notificationsSlice'
 import { useAppDispatch } from '@/store'
 import useChains from '@/hooks/useChains'
 import { Check, X } from 'lucide-react'
-import PaginatedDataTable, { type DataTableColumn } from '../PaginatedDataTable'
+import PaginatedDataTable, { type DataTableColumn } from '@/components/common/PaginatedDataTable'
 import { cn } from '@/utils/cn'
 import AddressCell from './AddressCell'
