+++ apps/web/src/features/spaces/components/SpaceAddressBook/SpaceAddressBookTable.tsx
@@ -11,7 +11,7 @@ import LocalContactActions from './LocalContactActions'
 import { formatDate } from '@/features/spaces/utils'
 import InitialsAvatar from '@/components/common/InitialsAvatar'
 import { useMemberNameResolver } from '../../hooks/useMemberNameResolver'
-import PaginatedDataTable, { type DataTableColumn, type ColumnWidth } from '../PaginatedDataTable'
+import PaginatedDataTable, { type DataTableColumn, type ColumnWidth } from '@/components/common/PaginatedDataTable'
 import { cn } from '@/utils/cn'
 import AddressCell from './AddressCell'
